tests/requirements.txt
Django==5.1
djangorestframework==3.15.2
pytest==8.3.2
coverage==7.5.1